feat(counts): `bind_count` is a `FreeCount`, and the negative is refused where it arrives (U1 sites 4-5)
The two counts on the AST's binding constructs are now the carrier the port
already used for the same quantity one field over: `New.bind_count` and
`Receive.bind_count` are `FreeCount`s, a negative count is refused at both
proto ingress points (`receive_from_proto`, `new_from_proto`) exactly as
`ReceiveBind.free_count`/`MatchCase.free_count` are (C52), and the `.max(0)`
pair in `well_scoped_receive`/`well_scoped_new` is deleted — there is nothing
left for it to defend against.

Why the count and not just the ingress. A negative `bindCount` is a value no
rule defines: it is a `well_scoped_par` depth, a sort-key leaf
(`sorter.rs`'s `leaf_i64`), the extent of `env.shift` in substitution, and the
argument to `new_bindings_cost`. The clamp hid the malformed message and left a
term whose sort key carried a value no rule defines. This is **stricter than
the Scala**, which carries a signed `Int` off the proto and validates nothing,
so it is a §6 deviation, registered beside the `BindPattern.freeCount` row; and
it is not a hard fork, because the counts the normalizer derives go through
`checked_level_count`.

`FreeCount::from_nonneg` is **deleted**. Its `debug_assert!(n >= 0)` was
compiled out in release, so the invalid count crossed silently in the builds
that matter. The carrier's total entries are now `from_len(usize)` (the sign is
not writable; the only escape is width, and it saturates rather than wrapping —
`shared/src/refined.rs`'s `BlockHeight`/`SeqNum` make the same choice for the
same reason) and the carrier's own `Add`. `count_free_vars` gained a typed
sibling, `count_free_vars_refined`.

Both directions of the site-5 claim, run rather than asserted:
- with the carrier, a probe building a count from a negative **does not
  compile**: `FreeCount::from_len(-1)` -> `error[E0600]: cannot apply unary
  operator `-` to type `usize``; `FreeCount(-1)` -> `error[E0603]: tuple struct
  constructor `FreeCount` is private`.
- with the old shape restored inside the module, the same probe compiles and in
  a **debug** test the assert fires (`free-count must be non-negative`, FAILED)
  while in a **release** test it passes — `i32::from(probe_old_shape(-1)) == -1`
  — i.e. the negative was silently accepted. That is the hole; the carrier is
  what closes it.

Falsified first for the wire: `a_negative_bind_count_is_refused_at_the_wire_boundary`
failed against the old pass-through (the decoders answered `Ok` with a negative
count in the term).

Deliberately left, with the reason in the code and in C52's row:
- `storage_printer`'s two reads of a **stored** `i32` field: `to_receive` now
  returns `Option<Par>` and a negative field refuses the row through the
  module's own fixed diagnostic message (`MALFORMED_PATTERN`) rather than
  clamping it to zero (a `for` binding nothing is a *different* term) or
  recovering a derivation — the port's own count convention cannot reproduce
  the value exactly, measured: for `for (@[x, ...rest] <- c)` the normalizer
  writes 2, the walk sees 1.
- `BindPattern.free_count` stays an `i32` and stays the durable fix C52 names:
  casper, node and the bench build that struct with literals, so the carrier
  cannot move onto the field from this crate. Both casper patterns are single
  free variables, so the convention gap above is latent there, not live.

Scoped: models 155, crypto 89, shared 80, rspace 170, rholang 226, zero
failures; clippy clean with the CI allowance list; `cargo fmt --check` clean for
these files; `tools/audit-type-system.sh` green on all four hard classes;
`tools/audit-test-register.sh` green (the count table refreshed to the tree's
live numbers, including the three rows other writers' tests had moved).

+++ rholang/src/reduce.rs
@@ -1887,13 +1887,19 @@ fn resolve_receive(
                     .map(|p| SortedProc::new(p.eval()))
                     .collect(),
                 remainder: rb.remainder.as_deref().cloned(),
+                // `BindPattern.free_count` is still an `i32` (casper/node construct the struct with
+                // literals), so the carrier is discharged here — the one boundary on this path.
                 free_count: i32::from(rb.free_count),
             },
             SortedProc::new(q),
         ));
     }
-    let subst_body =
-        substitute_par_and_charge(&receive.body, 0, &env.shift(receive.bind_count), cost)?;
+    let subst_body = substitute_par_and_charge(
+        &receive.body,
+        0,
+        &env.shift(i32::from(receive.bind_count)),
+        cost,
+    )?;
     Ok(Effect::Consume(
         binds,
         ParWithRandom {
@@ -1934,10 +1940,12 @@ fn resolve_new(
     urn_map: &BTreeMap<String, Par>,
     cost: &CostAccounting,
 ) -> Result<Effect, RholangError> {
-    cost.charge(Costs::new_bindings_cost(new.bind_count as i64))?;
+    cost.charge(Costs::new_bindings_cost(i64::from(i32::from(
+        new.bind_count,
+    ))))?;
     let mut r = (*rand).clone();
     let new_env = alloc(
-        new.bind_count,
+        i32::from(new.bind_count),
         &new.uri,
         &new.injections,
         env,
